Multi-link portals
This pull request adds multi-link option.

using "null" as a player when creating a link is meant to work like it worked before.
using an actual player should create a private link for them that no one else should see

THIS CODE SHOULD NOT BE USED (obviously). It's just a proof of concept.
Sorry if I am wasting your time